Removed useless installations.
Few people will want mercurial, and a more up to date version can
be installed from pypi. Nobody needs ssh anymore to get into
containers, but if necessary it can be installed by the user. pip
and setuptools will already be up to date, so there is no need to
install them.